setup-timezone: refactor. -i for install all tzdata and -k for keep copies
By default we copy the tzdata file to /etc/zoneinfo and symlink to our
copy. We also purge all previously copied timezones unless -k is given.

If -i is specified, then will the tzdata apk package stay and the
symlink will point to /usr/share/zoneinfo/... instead of a copy in
/etc/zoneinfo/