Modify setup-disk to correctly ignore swap when appropriate.
Simply adding a '0M,82' to sfdisk causes it to create a single-sector swap partition as opposed to no partition at all. In spirit of adhering to both theprinciple of least surprise *and* sensible partitioning, this case statement ensures that if swap is set to zero, the partition is never created at all.

Common SSD VPS providers encourage customers to not enable swap whatsoever. This will assist their template generation.