setup-user: refactor. ask for username first
The username is shorter than full name so ask for that first and suggest
the username as fullname which will make the username show up in login
menu instead of 'Linux User'.

If -f FULLNAME is specified, then give a decent default suggestion based
on first letter in first name and lastname.