Add unlock argument for setup-user
Adds a -u argument: useful for if you want to create a user non-interactively that logs in with an ssh key (Passed by -k) without needing to unlock the user manually afterwards.

The reasoning for this is that if you set no password needed, adduser automatically locks the account from logging in.

Empty SSH logins are already prevented, so this means a user can log in using ssh with the already configured key and simply use passwd when logging in with no other changes if -u is passed.

I pipe the stderr to /dev/null as regardless of whether the unlock is needed / does anything, passwd reports that the password was reset or if the user was already unlocked: I presume that, similar to adding groups, the user knows what changes they are requesting. It may be preferred to clear $forceunlock if the credentials are being added interactively: I prefer just performing the action regardless, but it may be desirable to report any odd errors besides the (2?) I assume users are going to see.