Add -u unlock command to setup-user
Copied from github PR, but add -u so that you can include to forcably unlock the user after creation.

Useful for when you want to create a user automatically so they can log in using a passed SSH key (-k), but without a password so they can set up a password after connecting.