setup-proxy: fix empty proxy
the while loop loops until the value of proxyurl is none or a http*://
scheme. however, when proxyurl is set to 'none', it then gets unset here
(`proxyurl= ;`), but the following `if` block checks if it is 'none' and
fails, so we configure an empty http_proxy= in the proxy.sh profile
script. this is then read by some software, and it causes it to crash.
for instance, flatpak via glib reads `http_proxy` from the environment,
and an empty variable is illegal.