setup-devd: also add hwdrivers for mdevd
The hwdriver service is currently activated for mdev and deactivated
when mdev is disabled. Since mdevd and mdev use (more or less) the
same mechanism, we also enable hwdrivers with mdevd.

The service is only started when a coldplug is required, because
if no coldplug is required, the appropriate kernel modules are
already loaded.

Note that with the MODALIAS line in /etc/mdev.conf, it is possible
that the hwdrivers service is entirely redundant. This would need
some additional testing.

Signed-off-by: Laurent Bercot <ska@appnovation.com>