setup-xorg-base: fix case with several sources in repositories
setup-xorg-base didn't correctly handle the case where
/etc/apk/repositories contains several main sources, and that
caused test failures. We fix this by throwing out the big sed command
(which fails on multiline /etc/apk/repositories) and processing the
repositories file in a pedestrian way, line by line in shell. Since
we entirely cut grep and sed and make sure we never fork in the loop,
the performance is acceptable.

Signed-off-by: Laurent Bercot <ska@appnovation.com>