setup-interfaces: find essids when SSID List capability present
When a essid has the 'SSID List' extended capibility, the `find_essids`
function will not return that ssid. That's because the `$1 ~ /SSID/` awk
pattern will match that line. The ssid variable then is set to an empty
value because there is no `:` (field separate) in that line, so `$2` is
empty.

The terminator of the statemachine then skips this essid, because it
checks if the `ssid` variable is not empty.

We can fix this by making sure the field ends with `SSID`, which is not
the case with the `SSID List` line.

Partial output of `iw dev wlan0 scan`:

```
        Extended capabilities:
                 * Extended Channel Switching
                 * TFS
                 * WNM-Sleep Mode
                 * TIM Broadcast
                 * BSS Transition
                 * SSID List
                 * Operating Mode Notification
                 * 6
                 * Max Number Of MSDUs In A-MSDU is unlimited
```

Reported-by: invoked
Fixes: #10546