setup-interfaces: find essids when UTF-8 SSID capability present
When a essid has the 'UTF-8 SSID' extended capibility, the `find_essids`
function will not return that ssid. This bug is very similar to the one
fixed in commit 6c9955e, and solved by ensuring nothing but whitespace
precedes the matched key in the regex.

Partial output of `iw dev wlan0 scan`:

```
        Extended capabilities:
                 * Extended Channel Switching
                 * UTF-8 SSID
                 * Operating Mode Notification
```