tests: setup-interfaces: more tests
Test that auto single eth0 interfaces creates the excpected
/etc/network/interfaces file.

Test that auto pick the first interface (eth0) when none are "up".

Test that auto set the link up for all interfaces, and down for all
except the selected interface. Test also that we pick the first that is
up.

Test that we don't set up dhcp on lo if there are no eth* interfaces.