update-kernel: fix handling of compressed firmware
Since the change to zstd compressed firmware[0] update-kernel missed most
of the firmware when creating the modloop. Now firmware should be included
regardless of compression.

I have tried if decompressing the firmware for the modloop improves the
size but it doesn't. The modloop was a little bit larger and a generated
iso was exactly the same size.

Fixes https://gitlab.alpinelinux.org/alpine/alpine-conf/-/issues/10594
Fixes https://gitlab.alpinelinux.org/alpine/aports/-/issues/16601
Closes https://gitlab.alpinelinux.org/alpine/aports/-/issues/16607
Closes https://gitlab.alpinelinux.org/alpine/alpine-conf/-/merge_requests/215
Closes https://gitlab.alpinelinux.org/alpine/alpine-conf/-/merge_requests/217

[0]: https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/60282